refactor(storage): publish the lane forwarder honestly
